refractor: fixed dart anti-patterns > > updated your file names(didn't change the name) to use recommended snake case by Dart
