+++ lib/UI/auth/details/movie_details_page.dart
@@ -1,22 +1,19 @@
-import 'dart:io';
-
 import 'package:cached_network_image/cached_network_image.dart';
 import 'package:carousel_slider/carousel_slider.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:movi/Repository/FirebaseRepository.dart';
-import 'package:movi/UI/Home/GenreMovies.dart';
-import 'package:movi/UI/Home/KeywordMovies.dart';
-import 'package:movi/UI/Home/CastPage.dart';
-import 'package:movi/UI/Widget/DrawerWidget.dart';
-import 'package:movi/UI/Widget/MovieWidget.dart';
-import 'package:movi/UI/Widget/TitleWidget.dart';
-import 'package:movi/Util/Responsive.dart';
-import 'package:movi/blocs/Casts_bloc/casts_bloc.dart';
-import 'package:movi/blocs/Movie%20Trailer/trailer_bloc.dart';
+import 'package:movi/UI/home/genre_movies.dart';
+import 'package:movi/UI/home/keyword_movies.dart';
+import 'package:movi/UI/home/cast_page.dart';
+import 'package:movi/UI/widget/drawer_widget.dart';
+import 'package:movi/UI/widget/movie_widget.dart';
 import 'package:movi/blocs/movieDetails_bloc/movie_details_bloc.dart';
+import 'package:movi/blocs/movie_trailer/trailer_bloc.dart';
+import 'package:movi/repository/firebase_repository.dart';
+import 'package:movi/util/responsive.dart';
+import 'package:movi/blocs/Casts_bloc/casts_bloc.dart';
 import 'package:movi/blocs/similar_bloc/similar_bloc.dart';
 import 'package:movi/models/UserMovies.dart';
 import 'package:percent_indicator/circular_percent_indicator.dart';
