+++ lib/UI/home/keyword_movies.dart
@@ -1,33 +1,35 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:movi/UI/Widget/MovieWidget.dart';
-import 'package:movi/Util/Responsive.dart';
-import 'package:movi/blocs/Homebloc/home_bloc.dart';
+import 'package:movi/UI/widget/movie_widget.dart';
+import 'package:movi/blocs/genre_movies/genre_bloc.dart';
+import 'package:movi/util/responsive.dart';
 import 'package:pagination_flutter/pagination.dart';
 
-class ViewAllTrending extends StatefulWidget {
-  const ViewAllTrending({
-    super.key,
-  });
+class KeywordMovies extends StatefulWidget {
+  final int keywordId;
+  final String genreName;
+  const KeywordMovies(
+      {super.key, required this.keywordId, required this.genreName});
 
   @override
-  State<ViewAllTrending> createState() => _ViewAllTrendingState();
+  State<KeywordMovies> createState() => _KeywordMoviesState();
 }
 
-class _ViewAllTrendingState extends State<ViewAllTrending> {
+class _KeywordMoviesState extends State<KeywordMovies> {
   int _current = 1;
 
   @override
   Widget build(BuildContext context) {
     return BlocProvider(
-      create: (context) => HomeBloc()..add(GetTrending(page: _current)),
+      create: (context) => GenreBloc()
+        ..add(MovieKeywordsFetched(page: _current, genreId: widget.keywordId)),
       child: Builder(builder: (context) {
         return Scaffold(
           appBar: AppBar(
             elevation: 0,
             backgroundColor: Colors.black54,
-            title: const Text("Trending",
-                style: TextStyle(
+            title: Text(widget.genreName,
+                style: const TextStyle(
                     color: Colors.white,
                     fontSize: 20,
                     fontWeight: FontWeight.w500)),
@@ -40,8 +42,8 @@ class _ViewAllTrendingState extends State<ViewAllTrending> {
             onPageChanged: (page) {
               setState(() {
                 _current = page;
-                BlocProvider.of<HomeBloc>(context)
-                    .add(GetTrending(page: _current));
+                BlocProvider.of<GenreBloc>(context).add(MovieKeywordsFetched(
+                    page: _current, genreId: widget.keywordId));
               });
             },
             nextIcon: const Icon(
@@ -80,15 +82,15 @@ class _ViewAllTrendingState extends State<ViewAllTrending> {
           body: SafeArea(
             child: Padding(
               padding: const EdgeInsets.symmetric(horizontal: 5),
-              child: BlocBuilder<HomeBloc, HomeState>(
+              child: BlocBuilder<GenreBloc, GenreState>(
                 builder: (_, state) {
-                  if (state is HomeLoading) {
+                  if (state is KeywordsLoading) {
                     return const Center(child: CircularProgressIndicator());
-                  } else if (state is TrendingLoaded) {
+                  } else if (state is KeywordsLoaded) {
                     return GridView.builder(
                         gridDelegate: SliverGridDelegateWithFixedCrossAxisCount(
                             crossAxisCount: Responsive.isDesktop(context)
-                                ? 9
+                                ? 8
                                 : Responsive.isTablet(context)
                                     ? 7
                                     : Responsive.isMobile(context)
@@ -97,24 +99,24 @@ class _ViewAllTrendingState extends State<ViewAllTrending> {
                             childAspectRatio: Responsive.isTablet(context)
                                 ? 0.5
                                 : Responsive.isDesktop(context)
-                                    ? 0.57
+                                    ? 0.65
                                     : Responsive.isMobile(context)
                                         ? 0.58
                                         : 0.58),
-                        itemCount: state.trending.results.length,
+                        itemCount: state.keywords.results.length,
                         itemBuilder: (_, index) {
                           return MoviesWidget(
-                              title: state.trending.results[index].title ?? "",
+                              title: state.keywords.results[index].title ?? "",
                               posterPath:
                                   // ignore: prefer_interpolation_to_compose_strings
                                   "https://image.tmdb.org/t/p/w500" +
-                                      (state.trending.results[index]
+                                      (state.keywords.results[index]
                                               .posterPath ??
-                                          state.trending.results[index + 4]
+                                          state.keywords.results[index + 4]
                                               .posterPath!),
-                              id: state.trending.results[index].id);
+                              id: state.keywords.results[index].id);
                         });
-                  } else if (state is HomeError) {
+                  } else if (state is KeywordsError) {
                     return Center(
                       child: Text(
                         state.message,
