+++ lib/UI/home/view_all/popular_movies.dart
@@ -1,9 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:movi/UI/Widget/MovieWidget.dart';
-import 'package:movi/Util/Responsive.dart';
-import 'package:movi/blocs/GenreMovies/genre_bloc.dart';
-import 'package:movi/blocs/Homebloc/home_bloc.dart';
+import 'package:movi/UI/widget/movie_widget.dart';
+import 'package:movi/blocs/home_bloc/home_bloc.dart';
+import 'package:movi/util/responsive.dart';
 import 'package:pagination_flutter/pagination.dart';
 
 class ViewAllPopular extends StatefulWidget {
