+++ lib/UI/widget/carousel_image.dart
@@ -1,6 +1,6 @@
 import 'package:cached_network_image/cached_network_image.dart';
 import 'package:flutter/material.dart';
-import 'package:movi/UI/Auth/Details/MovieDetailsPage.dart';
+import 'package:movi/UI/auth/Details/movie_details_page.dart';
 
 class CarouselImage extends StatelessWidget {
   final double width;
