+++ lib/UI/widget/movie_widget.dart
@@ -1,51 +1,60 @@
 import 'package:cached_network_image/cached_network_image.dart';
 import 'package:flutter/material.dart';
-import 'package:movi/UI/Auth/Details/Cast/CastWidgetPage.dart';
+import 'package:movi/UI/auth/Details/movie_details_page.dart';
+import 'package:movi/util/responsive.dart';
 
-class CastsWidget extends StatelessWidget {
-  final String? posterPath;
-  final String castId;
-  final String name;
-  final String role;
-  const CastsWidget(
-      {super.key,
-      required this.posterPath,
-      required this.castId,
-      required this.name,
-      required this.role});
+class MoviesWidget extends StatelessWidget {
+  const MoviesWidget({
+    super.key,
+    required this.title,
+    required this.posterPath,
+    required this.id,
+    this.voteAverage = 0.7,
+  });
+
+  final int id;
+  final String posterPath;
+  final String title;
+  final double? voteAverage;
 
   @override
   Widget build(BuildContext context) {
     return Padding(
-      padding: const EdgeInsets.only(
-        right: 10,
-      ),
+      padding: EdgeInsets.only(
+          right: Responsive.isDesktop(context) ? 20 : 5, top: 10, bottom: 2),
       child: SizedBox(
+        height: MediaQuery.of(context).size.width > 900
+            ? MediaQuery.of(context).size.height * 0.35
+            : MediaQuery.of(context).size.height * 0.25,
         width: MediaQuery.of(context).size.width > 900
-            ? MediaQuery.of(context).size.width * 0.07
-            : MediaQuery.of(context).size.width * 0.2,
+            ? MediaQuery.of(context).size.width * 0.12
+            : MediaQuery.of(context).size.width * 0.31,
         child: Column(
           mainAxisAlignment: MainAxisAlignment.start,
           crossAxisAlignment: CrossAxisAlignment.center,
           children: [
             GestureDetector(
               onTap: () {
-                debugPrint("cast id is $castId");
                 Navigator.of(context).push(MaterialPageRoute(
-                    builder: (context) => CastDetails(
-                          id: castId,
+                    builder: (context) => MovieDetails(
+                          id: id.toString(),
                         )));
+
+                debugPrint("the movie id is $id");
               },
               child: CachedNetworkImage(
-                height: MediaQuery.of(context).size.height > 400
-                    ? MediaQuery.of(context).size.height * 0.12
-                    : MediaQuery.of(context).size.height * 0.12,
-                imageUrl: "https://image.tmdb.org/t/p/w500$posterPath",
+                height: MediaQuery.of(context).size.width > 900
+                    ? MediaQuery.of(context).size.height * 0.3
+                    : MediaQuery.of(context).size.height * 0.2,
+                width: MediaQuery.of(context).size.width > 900
+                    ? MediaQuery.of(context).size.width * 0.12
+                    : MediaQuery.of(context).size.width * 0.3,
+                imageUrl: posterPath,
                 fit: BoxFit.cover,
                 imageBuilder: (context, imageProvider) {
                   return Container(
                     decoration: BoxDecoration(
-                      shape: BoxShape.circle,
+                      borderRadius: BorderRadius.circular(10),
                       image: DecorationImage(
                         image: imageProvider,
                         fit: BoxFit.cover,
@@ -66,28 +75,16 @@ class CastsWidget extends StatelessWidget {
               ),
             ),
             const SizedBox(
-                    height: 5,
-                  ),
+              height: 10,
+            ),
             Text(
-              name,
-              // overflow: TextOverflow.ellipsis,
+              title,
+              overflow: TextOverflow.clip,
               maxLines: 1,
               textAlign: TextAlign.center,
               style: const TextStyle(
-                  color: Color.fromARGB(255, 255, 255, 255),
+                  color: Colors.white,
                   fontSize: 15,
-                  fontWeight: FontWeight.bold),
-            ),
-            const SizedBox(
-              height: 1,
-            ),
-            Text(
-              role,
-              maxLines: 1,
-              textAlign: TextAlign.center,
-              style: const TextStyle(
-                  color: Color.fromARGB(255, 255, 255, 255),
-                  fontSize: 12,
                   fontWeight: FontWeight.bold),
             ),
           ],
