+++ lib/UI/widget/profile_card.dart
@@ -1,8 +1,8 @@
 import 'package:cached_network_image/cached_network_image.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
-import 'package:movi/UI/Home/ProfilePage.dart';
-import 'package:movi/Util/Responsive.dart';
+import 'package:movi/UI/home/profile_page.dart';
+import 'package:movi/util/responsive.dart';
 
 enum user { profile, logout }
