+++ lib/blocs/genre_movies/genre_state.dart
