+++ lib/blocs/movie_trailer/trailer_bloc.dart
