+++ lib/blocs/movie_trailer/trailer_state.dart
