+++ lib/blocs/search_movie/search_movie_bloc.dart
