+++ lib/blocs/user_casts_bloc/user_state.dart
